update list collection market collection detail hard and sort in filter and search

+++ lib/data/response/collection_detail/collection_detail_response.dart
@@ -63,7 +63,7 @@ class CollectionDetail {
   @JsonKey(name: 'nft_owner_count')
   int? nftOwnerCount;
   @JsonKey(name: 'total_volume_traded')
-  int? totalVolumeTraded;
+  double? totalVolumeTraded;
   @JsonKey(name: 'collection_standard')
   int? collectionStandard;
   @JsonKey(name: 'social_links')
